refactor(shared): file-organization sweep over app/shared/services and app/shared/util
The world's initial-metadata shape is a model beside the world data, and the two achievement definition helpers move under definitions/, the one subfolder that reads them. The relative imports in ImageSourceWhitelist stay: nuxt.config evaluates it before the aliases exist, and the file says so.
