+++ apps/web/shared/models/dungeons/data/world/InitialMetadata.ts
@@ -1,0 +1,7 @@
+import type { Direction } from "#shared/models/dungeons/data/player/Direction";
+import type { Position } from "grid-engine";
+
+export interface InitialMetadata {
+  direction: Direction;
+  position: Position;
+}
