+++ apps/web/shared/services/achievement/definitions/WebpageAchievementDefinitionMap.ts
@@ -1,6 +1,6 @@
 import { AchievementCategory } from "#shared/models/achievement/AchievementCategory";
-import { defineAchievementDefinition } from "#shared/services/achievement/defineAchievementDefinition";
-import { defineAchievementDefinitionMap } from "#shared/services/achievement/defineAchievementDefinitionMap";
+import { defineAchievementDefinition } from "#shared/services/achievement/definitions/defineAchievementDefinition";
+import { defineAchievementDefinitionMap } from "#shared/services/achievement/definitions/defineAchievementDefinitionMap";
 import { WebpageAchievementName } from "@esposter/db-schema";
 
 export const WebpageAchievementDefinitionMap = defineAchievementDefinitionMap(AchievementCategory.Webpage, {
