+++ apps/web/shared/services/dungeons/scene/world/getInitialMetadata.ts
@@ -1,14 +1,9 @@
-import type { Position } from "grid-engine";
+import type { InitialMetadata } from "#shared/models/dungeons/data/world/InitialMetadata";
 import type { SetOptional } from "type-fest";
 
 import { TilemapKey } from "#shared/generated/tiled/propertyTypes/enum/TilemapKey";
 import { Direction } from "#shared/models/dungeons/data/player/Direction";
 import { NotFoundError } from "@esposter/shared";
-
-interface InitialMetadata {
-  direction: Direction;
-  position: Position;
-}
 
 const TilemapInitialMetadataMap: Partial<Record<TilemapKey, SetOptional<InitialMetadata, "direction">>> = {
   [TilemapKey.Home]: { position: { x: 6, y: 21 } },
